Use the correct fd command
On Debian based system the fd executable is named 'fdfind'. This is due
to a possible name clash with an executable from the package 'FDclone'.

First try whether 'fdfind' is available. If not check that the command
'fd' is really the file finder.

Also update github workflow to install the fd-find package.