Move index files to ~/.qc/index
The managed index files are now created in ~/.qc/index, to keep them
separated from scripts and config. Also the log file from the cronjob
is created in ~/.qc/index.
The manual index index.dstore stays in ~/.qc.

Introduce new variables for directories for index, config and data, to
prepare for a later reorg (maybe XDG).